able to run with comments

+++ PanelQR1/TallShinnyQR.h
@@ -9,10 +9,10 @@
 
 // 注意M必须<=256,N必须<=32
 // 另外n必须<=N
-template <typename T, size_t M, size_t N>
-void hou_tsqr_panel(cublasHandle_t cublas_handle, size_t m, size_t n, T *A,
-                    size_t lda, T *Y, size_t ldy, T *R, size_t ldr, T *work,
-                    size_t lwork) {
+template <typename T, int M, int N>
+void hou_tsqr_panel(cublasHandle_t cublas_handle, int m, int n, T *A,
+                    int lda, T *Y, int ldy, T *R, int ldr, T *work,
+                    int lwork) {
     if (n > N) {
         std::cout << "hou_tsqr_panel QR the n must <= N" << std::endl;
         exit(1);
@@ -30,15 +30,15 @@ void hou_tsqr_panel(cublasHandle_t cublas_handle, size_t m, size_t n, T *A,
 
     // 2.使用按列进行分段的方式进行QR分解
     // 2.1 把瘦高矩阵进行按列分段
-    unsigned int blockNum = (m + M - 1) / M;
+    int blockNum = (m + M - 1) / M;
 
     // 2.2直接创建这么多个核函数进行QR分解,A中存放Q, work中存放R
-    my_hou_kernel<T, M, N><<<blockNum, blockDim>>>(m, n, A, lda, Y, ldy, R, ldr,
-                                               work, lwork);
+    my_hou_kernel<T, M, N>
+        <<<blockNum, blockDim>>>(m, n, A, lda, Y, ldy, R, ldr, work, lwork);
 }
 
 template void hou_tsqr_panel<float, 128, 32>(cublasHandle_t cublas_handle,
-                                             size_t m, size_t n, float *A,
-                                             size_t lda, float *Y, size_t ldy,
-                                             float *R, size_t ldr, float *work,
-                                             size_t lwork);
+                                             int m, int n, float *A,
+                                             int lda, float *Y, int ldy,
+                                             float *R, int ldr, float *work,
+                                             int lwork);
